replace isl_union_map_compute_flow by isl_union_access_info_compute_flow
The original interface was error-prone and unextendable.
In particular, the user could easily confuse the different input
and output arguments and the number of inputs and output was fixed.
The most pressing need for allowing other inputs is that we want
the user to be able to pass a schedule tree instead of a schedule map,
without also forcing the user to switch to schedule trees.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 6092abe194dbec6a5f4ba985ecaaea211672679a