isl_union_access_info: keep track of schedule tree
That is, rather than immediately converting the schedule tree into
a schedule map, we now store it in the isl_union_access_info.
We still convert it to a schedule map before we actually
initiate the dependence analysis, but this will be changed
in the next commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 61eb0d5992006704b8e7484e30d9f7975c6ffbe3