update for rename of FunctionProtoType accessors from 'Arg' to 'Param'
We usually use the new name in the code and map that to the old name
for older versions, but in this case, the new method name is also
used in other objects and those names should not be mapped to the old name.
It is therefore easier to keep using the old name and to map that to
the new name.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 0595033e1a811ae15b2d675190d544a3e73383b2