isl_pw_*_gist: special case context equal to single domain
If the context is equal to one of the domains, then the result
of the gist is simply the corresponding function on the universe domain.
The default implementation is not able to detect this case
when the context is a disjunction.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 17279c1d5e962e9863a6e72138026ee475372caa