isl_ast_codegen.c: generate_domain: intersect with simple hull of build domain
We used to intersect with the build domain itself, but this domain
may be a disjunction and intersecting with it may result in a lot of
extra conditions because it is more difficult to simplify away
a disjunction.  At least, isl's gist is not very good at simplifying
away disjunctions at the moment.
Note that by approximating the build domain, we may lose some special
cases where we would detect single-valuedness,
but we can always recover from that through the nested call.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 0ed6ed381cd8d743e6247cc7a9f077257be03bee