isl_basic_map_output_defining_equality: accept divs that do not depend on output
Originally, we would not consider an equality to be a defining equality
if it depends on any integer divisions.  However, if it only depends
on integer divisions that are defined in terms of parameters and
input dimension, then we can also consider the equality to be
a defining equality.

This allows us to recover an affine expression from
the set representation of an expression such as

	[N] -> { [3*floor(N/2) + 5*floor(N/3)] }

Without these changes, we would resort to performing
parametric integer programming to recover a piecewise
affine expression, which could (and in this case does)
involve multiple pieces.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 8f144dc408423d131e3beb30472d5061d426b4c6