document isl_space_tuple_match as isl_space_tuple_is_equal
The function isl_space_tuple_match is used by the AST generator,
so it should be part of the public API.  Rename it to
isl_space_tuple_is_equal, because "match" is not a very clear
description.

The old name is kept for now in case it was already being used.
It will be removed at some point.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: c73f000ccf10ec381b99498b599dc38a27508151