isl_union_*_add_*: perform check for "zero" part unconditionally
Inside isl_union_*_add_*, we only want to check if the part
can be dropped, not whether it is actually equal to zero.
We therefore do not need to check if an empty piecewise expression
means that it is equal to zero.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 4644e040e9074fa206013d111e169d88be790dd9