isl_ast_build.c: update_values: explicitly take into account stride constraint
Before this commit we would rely on the caller to add the stride constraint
to the build domain.  However, we want to avoid adding this stride constraint
to the build domain in some cases and so we have to take it into account
explicitly while checking if the dimension attains a single value.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 950d2defe302f3f1857831676310a4355708170d