isl_basic_map_gist: avoid (temporarily) adding context constraints to input
We add the context constraints to the input in order to be able
to compute the affine hull of the intersection.
However, there is no need to use this intersection for the actual
gist computation.  Doing so only risks replacing constraints
in the input by constraints in the context in case of mutuallly
redundant constraints.

Simply use the original input to perform the actual gist computation.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: e72d7e21dd8a22f65b00acd847a8e795d7a2c0d6