AST generator: exploit implied guard to simplify inner guards
Since we have to generate the implied guard any way, we might as well
use it to simplify the guards in the children.
It is not clear if this would ever have any effect in practice.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: f4485101547c8e15de093ba1d02757b9f174dd02