isl_map_coalesce: check for subsets with smaller number of divs
If the divs in the pair of basic maps we are trying to coalesce
are not the same, then we already check whether one is a subset of
the other, but only in cases where the potential subset has more divs
than the potential superset.
However, it is also possible for a subset to have fewer divs.
In particular, if the subset satisfies an extra equality,
then some of the divs may have been simplified away.
To detect these cases, we apply the extra equalities of the
possible subset to the possible superset (simplifying its divs)
and then check if the first is indeed a subset of the intersection.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 06e7081fc61b76ce27dc6f5e44750efbc718262b