isl_coalesce.c: coalesce: keep track of how a pair of basic maps is combined
This allows us to avoid rechecking pairs of basic maps that have
been considered before when any coalescing has been performed,
based on the way the two basic maps were coalesced.
In particular, the two basic maps may have been fused together or
one of the basic maps may have been dropped because it is a subset
of the other.  When the basic map that is compared against other
basic maps is dropped, then we can move on to a new basic map
and compare that to the already considered basic maps.
When the other basic map is dropped, then we continue comparing
the current basic map against the other basic maps.
Only when the two are fused do we need to start over with the fused
basic map.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 26b1fd70cdbd1a4d8b0bb55daffd5d4bf15715d4