isl_coalesce.c: set_from_updated_bmap: update underlying set
If we first update the basic map and then obtain the underlying set,
then the integer divisions could get dropped in case the tableau
turns out to be empty.
It's not clear if this can ever happen in the current state of
the code, but it is better to be safe.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: 04e30b749a2a963ef60ff07fbde27172ba190705