add isl_tab_insert_var
This allows us to insert a variable in the tableau at a specific position,
rather than at the end.  This also means that during an undo, we may have
to remove a variable that is not in the last position so that drop_col
needs to be modified accordingly.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: c55779905d0992081a7cf6f92f0f966e95266a55