isl_map_coalesce: remove hidden factors from constraint coefficients
Removing these hidden factors improves the chances of an inequality
being recognized as adjacent to an equality.

For example, the set

	{ [x,y] : 2x = 3y and 0 <= y <= 4; [-3,-2] }

would not get coalesced before because the inequality 0 <= y
is not recognized as being adjacent to an equality of the second
basic map.  After removing common factors, the set looks like

	{ [x,y] : 2x = 3y and 0 <= x - y <= 2; [-3,-2] }

and now the constraint 0 <= x - y is recognized as being
adjacent to an equality of the second basic map.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>

Original commit from http://repo.or.cz/w/isl.git: a929cb74cb3eeaa3200f266fde48259b781baa85